Add return values to ModelActions.call
Adds return values to all ModelActions.call overloads that can supply one.
This is accomplished primarily by two changes:
1. We add a `writingTo` method to Task, which produces a new task which will hand its return value to a Consumer when it completes.
2. In both ReplayingReactionContext and ThreadedReactionContext, we add a MutableObject to collect the return value, and ask the child task to write its return value there.

Since the TaskHandle ensures that the `call` in each Context only returns if the child task completes, we can be assured that the return value is populated before it is read.

I think that collecting the return value at this step minimizes the number of places we duplicate the "collect a return value" logic, though I could be wrong about that.

It's also possible we could do this without the side-channel of a MutableObject, which would be cleaner in some ways, but I think that would require more extensive changes to the engine.