Debug call return values for replaying tasks
Debugs getting a return value for `call` from a replaying task.
Using the `Supplier` overload of `Memory.doOnce` doesn't work, because that method treats the task as a "read", and records it as done when the value is returned.
Since the tasks throws an exception to yield before returning its value, it's never treated as "done", so the child task is dispatched over and over forever.

To fix this, we ask the child task to write its return value directly to replaying task memory when it finishes.
Then, we perform a "faux read" to retrieve that value and update the memory object's bookkeeping.
Reaching into the Memory object directly like this doesn't feel particularly clean, but it works.
Maybe this should be promoted to a supported method on the memory object?
The other side, using a "read" operation to retrieve the returned value, seems fine to me.