Add and debug ground events
Adds the ability to model ground events, including an AuxiliaryModeling ground event.
That event allows you to inject arbitrary modeling behavior from an activity definition into a sequence, to more expressively define "deep" modeled activities.
To demonstrate these new events, Do_Observation_Deep was cloned, and the new events used to inject the stand-ins for additional activity-level modeling.

Note that the AuxiliaryModeling ground event is different from the GenericGroundEvent, which indicates a *lack* of modeling for an otherwise meaningful event.
By contrast, the AuxiliaryModeling ground event indicates the *presence* of additional modeling, despite a lack of otherwise-meaningful events.

For example, we might use the GenericGroundEvent to capture an azimuth crossing event which is meaningful to the DSN, but not to us.
We might use an AuxiliaryModeling event to update the "lastMaintenanceTime" state in our model from the DoMaintenance activity, if no one command is solely responsible for "maintenance".

Additional changes:
- Bug fix in the FcplSequenceEngine for zero-time commands where there were collisions on the command complete flag.
- More Jackson-idiomatic polymorphic deserialization of SeqJSON