WIP - Incon behaviors
Adds the InconBehavior interface and some of the surrounding machinery to connect it.

The thrust of this idea is to wire together incon reading, fincon writing, and cell creation,
such that the easiest way for a modeler to create a new cell is to correctly wire up the incons.
This means that wiring up those incons should be relatively painless,
it should minimize duplicating necessary information,
it should eliminate specifying any unnecessary information,
and side-stepping the incons process should require rebuilding a large portion of the resource stack, as a strong discouragement to doing so.

This commit does not compile. The status of various classes is:
1. I'm mostly happy with InconBehavior itself. It's a neatly packaged bit of behavior, it composes well (see InconBehavior.map), and I think it's legible as-is.
2. I may want to remove InconBehaviors static constructors "constant" and "serialized". I think those belong in MutableResource, as part of the "notSaving" and "serializing" helper methods instead.
3. I'm not happy with MutableResource yet, as we're now duplicating information for the incon behavior (name and ValueMapper) that we'll give later if/when we register it.
   I'm wondering if we should roll registration into this method as well, perhaps with a flag or alternate constructor to not register a resource...
4. Everywhere we build MutableResources needs to be updated, pending what we decide to do for the step above.
   Additionally, we'll need to document how this changes the MutableResource interface and include that in the migration notes should this be adopted by Aerie.