Expose global absolute clock
Using the absolute-timed InstantClock, add a new global simulation clock created by StreamlineSystem.init.
Also, refactor the arguments to StreamlineSystem.init into an InitArgs object with an accompanying builder.
This lets us write a "testBuilder" constructor that fills in as many init args with defaults as it can, while letting us override parameters if needed.
This pattern will extend well if we need to add more parameters in the future.

Using the global simulation clock, removes the "plan start time" parameter from the overload of DiscreteResources.precomputed that uses Instant keys.
This guarantees that absolute-timed precomputed resources are synced with the global clock.
This also creates a name collision with the version of precomputed that uses Duration keys, so the one with Instant keys was renamed to precomputed$.