Bug fix: Consult clock for fixedTimeCondition
When doing clamped integration, we need to build fixed-time conditions. These conditions fire after a fixed time, computed in advance for performance.
Previously, the fixed time condition reported it would be satisfied in a fixed amount of time, regardless of when it was asked.
This is incorrect if the condition is queried at any time later than its creation; the condition appears to "retreat" in time as it's queried.

Fix this bug by having the fixedTimeCondition consult the clock, so it can factor in the time elapsed since its creation.