Change SimulationResultsWriter to write more directly to file
Changes SimulationResultsWriter to write output JSON more directly to the chosen output file (or stdout).
It no longer builds in-memory JsonValue nor String representations of large chunks of the output.
Instead, the resultsGenerator is connected directly to the chosen output stream, and build* subroutines to build in-memory JsonValue objects
are replaced by corresponding write* subroutines to write that section of the output directly to file.

When using a ResourceFileStreamer, this means opening only one resource temp file at a time and streaming data from that file to the final output file, never loading the full resource profile into memory.

This makes stateless Aerie capable of handling very large simulation results. This was tested with a Clipper plan that generated a nearly 14GB output file.